docs: clarify how to connect to saas cluster
It should be easy to connect the Zeebe REST Client to your own SaaS
cluster. So I've clarified this.

It does not yet clarify how to connect to a Self-Managed cluster, as
I've not yet been able to connect the Zeebe REST Client (running as
Docker container) to a locally running Zeebe Broker.

closes #1