test: get status should respond with topology
This commit adds a first test to verify the behavior of the controllers.
Sepcifically, on GET /status we expect that it responds with the
topology provided by the Zeebe Client.

A FakeZeebeClient is provided to act in behalf of the real ZeebeClient
in tests. This required a small amount of glue-code, but in the end
provides simple usage in tests.