test: support vars in post /process-instances
This tests should verify that the variables are really passed along to
the zeebe client, but it doesn't. At this time, I don't see an easy way
to test that. Instead, we can use test containers later to fully verify
the correct behavior of the rest api, connecting it to a real zeebe
container.