docs: explain how to connect to self-managed
If we want this tool to be easy to get started with, it should clearly
instruct how to connect to a Camunda Platform cluster. That also
includes how to connect to a self-managed cluster.

Finding the IP can be hard, and I was not yet able to clearly document
this. Often, it's best to just look it up in the logged configuration at
startup of the gateway (or broker with embedded gateway).

For the rest, this documentation focuses on dealing with secure
communication. I'm assuming here that secure connection is enabled on
the cluster they are trying to connect to, but if not then the
documentation expains how to disable it. I believe this is the most
secure way, and nudges users towards a secure solution.