feat: add status property to activated jobs
In order to fit job completion (and later failure) into the REST
methods, we'll need a way to say that a job is completed.

Typically, this is achieved by patching a specific job's status:
- `PATCH /jobs/123 {"status":"completed"}`.

However, that requires a job to have a `status` property, which it did
not have by default.

This commit adds a status property to jobs.