test: support PATCH /jobs/{key} to complete jobs
Adds the test for the PATCH /jobs/{key} endpoint, which can be used to
change the status of a job. This can effectively complete a job by
updating the status to `completed`.

Other status' are not yet supported, but we can imagine that failing a
job will work in the same way.

We can also imagine using this endpoint in the future to set variables
for a job's scope, if users would want something like that.