feat: support PATCH /jobs/{key} to complete jobs
Adds the endpoint to PATCH /jobs/{key}, which can be used to update a
job. Currently, it only allows updating the status of a job to
`completed`. This effectively completes the job in Zeebe.

In the future this endpoint may also be used for other job updates, e.g.
failing a job.