build: add openapi generator
Adds the dependencies and plugin necessary to generate code from an
openapi specification.

We want to generate code from a spec (design-first) rather than the
other way around (code-first), as this is considered a best practice by
the OpenAPI initiative.

A simple mustache template was added to resolve a problem with the
code generated by v6 of the openapi-generator, when used in combination
with the kotlin-spring and springdoc-openapi-ui. It generates one file
which is annotated with the jakarata variant of the @Generated
annotation. Since my SpringBoot version is not yet at v3+ we need to
adjust this generated code.