fix(api): rename retryBackOff to retryBackoff
The concept is wrongly named BackOff implying that the retry should back
off, but instead we're discussing a retry backoff (single word).

Note that this is also wrong in the actual Zeebe API (gRPC).
See: https://docs.camunda.io/docs/apis-clients/grpc/#failjob-rpc

However, I feel that this API should be clear, and it is allowed to have
differences from the gRPC API, because it is not the gRPC API itself.