fix: include all expected states in err msg
The `status` property of the PATCH /jobs/{key} can be one of:
- completed
- failed
- error_thrown

But, only `completed` was mentioned in the error message, when the
status is unknown.

This string became a bit long so I split it over 3 lines. However, it
then requires us to use extra methods like trimMargin and we need to
replace the new-line characters. I'm not sure whether this is really
better than a broken down string with `+` operator.