ci(deploy): use docker/login-action with pat
There was an issue with publishing the package to ghcr when dependabot
merged the pull request. I guess that the publish is denied (403)
because dependabot does not have rights to write this package.

I've added a PAT with `write:packages` as a secret to allow the
docker/login-action to login as my account. This should be safe.