docs(openapi): add GET /process-instances/{key}
This operation will allow to retrieve the details of a specific process
instance, like the current status, when it was started and when it was
ended if it has already ended.

A specific note must be made to consider the eventual consistency, as
this information will be retrieved from Operate which lags behind due
the architecture of the system. Users may need to request the
information multiple times to deal with this.