feat: add GET /process-instances/{key} operation
Adds the implementation for this operation. It uses Operate to retrieve
the data. This requires some configuration to connect to Operate.

Instead of writing my own configuration to connect to Operate, I copied
the implementation used by the connectors. It is likely that this
implementation will eventually move to spring-zeebe, at which point we
can remove the copied java classes. For now, these should provide
configurations to connect to Operate in SaaS, Self-Managed with
Keycloak, and Self-Managed using a simple username/password. The details
should be documented in the getting started / configuration.

Since the camunda-operate-client-java only provides java.util.time.Date
objects for the start and end datetime of the process instance, we need
to assume the timezone. We maybe able to deal with this in a more
sufisticated manner but this simple approach would provide value in most
common cases already.