fix(operate): allow missing configuration
The copied OperateCLientFactory and OperateClientLifecycle did not allow
for the configuration to be missing when the Lifecycle is injected.
However, in this project users should be free to configure Operate or
not.

If Operate is not configured, we should inform the client with a 503
Service Unavailable, similar to how the Zeebe connection may be left
unconfigered.

We can improve the user experience by clearly communicating an error
message or by logging the error. In the future, we can also add health
checks that show which connections are configured correctly and which
aren't.

By leaving Operate's configuration optional, we allow future community
support to other data sources like ZeeQS, and others.

The implementation of this change uses an in-method record definition,
as it is only used to store the result of both configuration getters:
url and auth.

In order to make the OperateClientLifecycle safe to use, we can simply
reuse the get() method instead of using delegate directly. This way an
exception is thrown for any request using the injected lifecycle if it
is not (or wrongly/partially) configured.