feat(docs): configure operate connection
This changes a lot of documentation because connecting to Operate can be
achieved in multiple ways, each with their own side-notes.

I've also chosen to add sections to these guides where users can check
whether the connection works as expected. And highlighted that they
should reach out if they get stuck. I'd rather not be complete in the
documentation, but have helpful error messages. If those messages are
not helpful enough yet then we need to improve them. The best way to
figure that out is through user feedback.

This commit also expands the configurations page by describing most of
the useful (and available) environment variables. What these do or how
they should be used, depends on whether the user is connecting to SaaS
or Self-Managed. I've added badges to highlight these. Note that the
environment variables are headings, so they can be linked to.

The custom css was messing up the look of the badge, so I disabled it.