feat(docker): add operate and elasticsearch
The default local setup should include Operate (and Elasticsearch) in
order to allow users to try out the Insights operations.

To connect to Operate, the REST API authenticates with the simple
username/password authentication on the demo account, which is available
by default.

Operate itself is available on port 8080, and locally forwarded from
8081. This makes it easy to connect to operate using http://operate:8080

The Elasticsearch service is configured using the example from
github.com/camunda/camunda-platform (docker-compose-core).