docs(README): add GET process-instances operation
This new operation requires a connection to Operate, which involved a
bunch of new documentation for the Connect to a cluster section.

Instead of copying that documentation into the README, it makes more
sense to guide users to the docs.

This commit therefore removes the connect to a cluster sections from the
README and replaces these with a Documentation section that refers users
to the different parts of the docs.

It also adds the Camunda Operate CLient as a new acknowledgement.